Move EVA protocol to ipv8 component
